# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baraka API",
        "description": "API backend pour la plateforme Baraka de microcrédit",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/baraka-backend/public",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/v1/portail/dashboard": {
            "get": {
                "tags": [
                    "Portail"
                ],
                "summary": "Tableau de bord de l'IMF connectée",
                "operationId": "55f8d00567d537aaa31b25d13f2ff1bb",
                "responses": {
                    "200": {
                        "description": "Données du tableau de bord"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/portail/stats": {
            "get": {
                "tags": [
                    "Portail"
                ],
                "summary": "Stats globales IMF sur 30j, détaillées par offre",
                "operationId": "2d1272cf1bcdee0745404612d91cb7cc",
                "responses": {
                    "200": {
                        "description": "Stats par offre"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/portail/stats/offre/{id}": {
            "get": {
                "tags": [
                    "Portail"
                ],
                "summary": "Stats d'une offre groupées par semaine sur 12 semaines",
                "operationId": "5daf808cec0b3e20bd52739f4ff06829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stats hebdomadaires"
                    },
                    "403": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/portail/stats/export": {
            "get": {
                "tags": [
                    "Portail"
                ],
                "summary": "Export CSV des stats sur 12 mois",
                "operationId": "69b966af74eae5c989dcae369346d537",
                "responses": {
                    "200": {
                        "description": "Fichier CSV"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/portail/profil": {
            "get": {
                "tags": [
                    "Portail"
                ],
                "summary": "Profil de l'IMF de l'agent connecté",
                "operationId": "05e9a9ab39f7b933af8cef6b4b9b08bf",
                "responses": {
                    "200": {
                        "description": "Profil IMF"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Portail"
                ],
                "summary": "Mise à jour du profil de l'IMF",
                "operationId": "bb19824c374c5c1417d3059c4cd66fd2",
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le token JWT obtenu via POST /api/auth/otp/verify"
            }
        }
    },
...
}